requirements.txt
kagglehub
datasets
pandas
requests
beautifulsoup4
boto3
python-dotenv